tidy: hard wrapping long lines in episode/SMP_for_research.md
Splitting this review into two to make it easier to see the changes, the first here mainly hard wraps long lines at 120
characters making it easier to find lines and in-keeping with the linting of `markdownlint-cli2`.

The next commit will contain edits/tpyos/tweaks although I already made a few changes before realising it would be
easier to make sense of the changes if split into to commits.