tidy: Use custom title in index.md; tidy a few other things
I actually went and read the commit that introduced the use of custom titles in `index.md` as I should have done in the
first instance and a tweak needs to be made to `config.yaml` (see
[342](https://github.com/carpentries/sandpaper/pull/342#issuecomment-1258738205)) which is included in this commit.

However, testing I found that we couldn't have just a YAML header in `index.md` this caused the error...

```
 ── Creating homepage ───────────────────────────────────────────────────────────
  Error in UseMethod("xml_find_first") :
    no applicable method for 'xml_find_first' applied to an object of class "xml_document"
```

I've therefore moved the opening paragraph from `learners/setup.md` > `index.md` and it seems to work.

Took the opportunity to improve a few other minor things (there was a complaint about the use of an uninformative `here`
as the text for a hyperlink).

This time I've checked _before_ making a pull request and the pages render as we would like.